% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.lmm.mu.R
\name{oss.lmm.mu}
\alias{oss.lmm.mu}
\title{Title}
\usage{
oss.lmm.mu(
  n,
  c,
  mu0 = 0,
  sig2mu = 100,
  a1 = 2.1,
  b1 = 4,
  a2 = 2.1,
  b2 = 4,
  lrep = 50,
  pmax = 500,
  pby = 100,
  prep = 3,
  nchains = 2,
  niter = 900,
  nburn = 300,
  nthin = 1,
  ncores = 1,
  plot = TRUE,
  ...
)
}
\arguments{
\item{n}{a}

\item{c}{a}

\item{mu0}{a}

\item{sig2mu}{a}

\item{a1}{a}

\item{b1}{a}

\item{a2}{a}

\item{b2}{a}

\item{lrep}{a}

\item{pmax}{a}

\item{pby}{a}

\item{prep}{a}

\item{plot}{a}

\item{...}{}

\item{chains}{a}

\item{iter}{a}
}
\value{
a
}
\description{
Title
}
\examples{
a
}
