% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osslmm-package.R
\docType{package}
\name{osslmm-package}
\alias{osslmm-package}
\alias{osslmm}
\title{The 'osslmm' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
