% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpos.lmm.R
\name{rpos.lmm}
\alias{rpos.lmm}
\title{Title}
\usage{
rpos.lmm(
  n,
  p,
  mu0 = 0,
  sig2mu = 100,
  a1 = 2.1,
  b1 = 4,
  a2 = 2.1,
  b2 = 4,
  chains = 2,
  iter = 100,
  nburn = 100
)
}
\arguments{
\item{n}{a}

\item{p}{a}

\item{mu0}{a}

\item{sig2mu}{a}

\item{a1}{a}

\item{b1}{a}

\item{a2}{a}

\item{b2}{a}

\item{chains}{a}

\item{iter}{a}

\item{nburn}{a}
}
\value{
a
}
\description{
Title
}
\examples{
a
}
